+++ Integrations/PhishLabsIOC_EIR/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-- change name to PhishLabsIOC_EIR_v2 from PhishLabsIOC_EIR
+
 
 ## [19.11.1] - 2019-11-26
 #### New Integration