+++ Integrations/PhishLabsIOC_EIR/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-
+- change display name phishlabs eir to phishlabs eir v2
 
 ## [19.11.1] - 2019-11-26
 #### New Integration