+++ Integrations/PhishLabsIOC_EIR/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+
+
+## [19.12.0] - 2019-12-10
 Changed the display name to **PhishLabs EIR v2**.
 
 ## [19.11.1] - 2019-11-26